Updating the selected date filter button to reflect if its after or before a certain date.
